refactor: merge dom.ts back into animation.ts
dom.ts functions (createGifPopup, createFallingItems, spawnAnimatedSprite)
are animation-related — the separate file was too granular. Merged back
into animation.ts with consolidated imports and tests.
