+++ src/themes/underwater.ts
@@ -1,7 +1,12 @@
 import type { Theme } from './types';
-import type { GifPopup } from '../utils/dom';
-import { createGlider, createDrifter, createAnimationLoop } from '../utils/animation';
-import { createGifPopup, spawnAnimatedSprite } from '../utils/dom';
+import type { GifPopup } from '../utils/animation';
+import {
+  createGlider,
+  createDrifter,
+  createAnimationLoop,
+  createGifPopup,
+  spawnAnimatedSprite,
+} from '../utils/animation';
 
 const THEME_ID = 'underwater';
 const THEME_SELECTOR = `[data-theme="${THEME_ID}"]`;
